Move assert_compare to test folder.
