+++ test/regression/road_curve/road_curve_test.cc
@@ -39,10 +39,10 @@
 #include <maliput/math/matrix.h>
 #include <maliput/math/vector.h>
 
+#include "assert_compare.h"
 #include "maliput_malidrive/road_curve/arc_ground_curve.h"
 #include "maliput_malidrive/road_curve/cubic_polynomial.h"
 #include "maliput_malidrive/road_curve/line_ground_curve.h"
-#include "maliput_malidrive/test_utilities/assert_compare.h"
 #include "maliput_malidrive/test_utilities/function_stub.h"
 #include "maliput_malidrive/test_utilities/ground_curve_stub.h"
